rename various common hooks
