+++ client/src/components/NewNote/useNewNote.ts
@@ -1,9 +1,9 @@
 import { useNewNoteMutation } from "@/lib/query/useNewNoteMutation";
 import useTagsQuery from "@/lib/query/useTagsQuery";
+import useAuthentication from "@/lib/useAuthentication";
 import useRouteProps from "@/lib/useRouteProps";
 import { queryClient } from "@lib/query-client";
 import { useTagSelection } from "@lib/state/selected-tags-state";
-import useAuthentication from "@lib/use-authentication";
 import type { NewNote } from "@type/server/note.types";
 import { useEffect, useState } from "react";
