+++ client/src/components/NewTag/useNewTag.ts
@@ -1,7 +1,7 @@
 import useTagsQuery from "@/lib/query/useTagsQuery";
+import useAuthentication from "@/lib/useAuthentication";
 import { queryClient } from "@lib/query-client";
 import { useTagSelection } from "@lib/state/selected-tags-state";
-import useAuthentication from "@lib/use-authentication";
 import type { NewTag } from "@type/server/tag.types";
 import { useEffect, useState } from "react";
 import { useNewTagMutation } from "./useNewTagMutation";
