+++ client/src/lib/useAuthentication.ts
