+++ client/src/lib/useReconcileSession.ts
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import useMeQuery from "./query/useMeQuery";
-import useAuthentication from "./use-authentication";
+import useAuthentication from "./useAuthentication";
 import { localUser } from "./user-storage";
 
 /** Reconcile client-side authentication state with server-side state. */
